requirements.txt
linkedin-api==2.2.0
PyYAML==6.0.2
requests==2.32.3
Unidecode==1.3.8
